+++ packages/vira/src/Vira/Web/Pages/PullPage.hs
@@ -8,10 +8,6 @@ GitHub check run creation is handled reactively by 'Vira.GitHub.Webhook'.
 module Vira.Web.Pages.PullPage (
   Routes (..),
   handlers,
-
-  -- * Shared widgets
-  forkBadge_,
-  prStateBadge_,
 ) where
 
 import Effectful (Eff)
@@ -27,7 +23,7 @@ import Vira.App qualified as App
 import Vira.App.CLI (WebSettings)
 import Vira.CI.Client qualified as Client
 import Vira.State.Acid qualified as St
-import Vira.State.Type (ForgeInfo (..), OwnerName (..), PRCommit (..), PRState (..), PullRequest (..), prBranchRef)
+import Vira.State.Type (ForgeInfo (..), PRCommit (..), PullRequest (..), prBranchRef)
 import Vira.State.Type qualified as St
 import Vira.Web.LinkTo.Type qualified as LinkTo
 import Vira.Web.Lucid (AppHtml, getLink, runAppHtml)
@@ -36,6 +32,7 @@ import Vira.Web.Widgets.Button qualified as W
 import Vira.Web.Widgets.Commit qualified as W
 import Vira.Web.Widgets.JobsListing qualified as W
 import Vira.Web.Widgets.Layout qualified as W
+import Vira.Web.Widgets.PullRequest qualified as W
 import Web.TablerIcons.Outline qualified as Icon
 
 data Routes mode = Routes
@@ -74,8 +71,8 @@ viewPRDetail pr commits jobs = do
     ( div_ [class_ "flex items-center space-x-3"] $ do
         div_ [class_ "flex items-center space-x-2 text-sm text-gray-600 dark:text-gray-300"] $ do
           span_ $ toHtml (unBranchName pr.headBranch) <> " → " <> toHtml (unBranchName pr.baseBranch)
-        forkBadge_ pr
-        prStateBadge_ pr.prState
+        W.forkBadge_ pr
+        W.prStateBadge_ pr.prState
         whenJust pr.forgeInfo $ \f ->
           a_ [href_ f.url, target_ "blank", class_ "text-gray-400 hover:text-gray-600 dark:hover:text-gray-300"] $
             div_ [class_ "w-5 h-5 flex items-center justify-center"] $
@@ -128,15 +125,6 @@ viewUnapprovedCommitRow pr pc =
 
 -- * UI Helpers
 
--- | Fork indicator badge (shows head owner when PR is from a fork)
-forkBadge_ :: (Monad m) => PullRequest -> HtmlT m ()
-forkBadge_ pr
-  | St.prIsFork pr =
-      span_ [class_ "inline-flex items-center px-2 py-0.5 rounded text-xs font-medium bg-purple-100 dark:bg-purple-900/30 text-purple-700 dark:text-purple-300"] $ do
-        div_ [class_ "w-3 h-3 mr-1 flex items-center justify-center"] $ toHtmlRaw Icon.git_fork
-        toHtml $ "fork: " <> unOwnerName pr.headOwner
-  | otherwise = mempty
-
 -- | Approve button for unapproved fork commits (posts to core approval route)
 approveButton_ :: RepoName -> Int -> CommitID -> AppHtml ()
 approveButton_ repoName prNum sha = do
@@ -147,21 +135,3 @@ approveButton_ repoName prNum sha = do
     $ do
       W.viraButtonIcon_ $ toHtmlRaw Icon.shield_check
       "Approve"
-
--- | Icon for PR state
-prStateIcon :: PRState -> ByteString
-prStateIcon = \case
-  PROpen -> Icon.git_pull_request
-  PRClosed -> Icon.git_pull_request_closed
-  PRMerged -> Icon.git_merge
-
--- | Badge for PR state
-prStateBadge_ :: (Monad m) => PRState -> HtmlT m ()
-prStateBadge_ prState = do
-  let (colorClass, label) = case prState of
-        PROpen -> ("bg-green-100 dark:bg-green-900/30 text-green-800 dark:text-green-300", "Open" :: Text)
-        PRClosed -> ("bg-red-100 dark:bg-red-900/30 text-red-800 dark:text-red-300", "Closed")
-        PRMerged -> ("bg-purple-100 dark:bg-purple-900/30 text-purple-800 dark:text-purple-300", "Merged")
-  span_ [class_ $ "inline-flex items-center px-2 py-0.5 rounded-full text-xs font-medium " <> colorClass] $ do
-    div_ [class_ "w-3 h-3 mr-1 flex items-center justify-center"] $ toHtmlRaw $ prStateIcon prState
-    toHtml label
