+++ packages/vira/src/Vira/Web/Widgets/JobsListing.hs
@@ -21,12 +21,13 @@ import Lucid
 import Lucid.Htmx.Contrib (hxPostSafe_)
 import Vira.App qualified as App
 import Vira.State.Acid qualified as St
-import Vira.State.Type (BranchBuildState (..), BuildFreshness (..), OwnerName (..), PRBuildState (..), PRCommit (..), PRDetails (..), PRState (..), PullRequest (..))
+import Vira.State.Type (BranchBuildState (..), BuildFreshness (..), PRBuildState (..), PRCommit (..), PRDetails (..), PullRequest (..))
 import Vira.State.Type qualified as St
 import Vira.Web.LinkTo.Type qualified as LinkTo
 import Vira.Web.Lucid (AppHtml, getLink, getLinkUrl)
 import Vira.Web.Widgets.Button qualified as W
 import Vira.Web.Widgets.Commit qualified as W
+import Vira.Web.Widgets.PullRequest qualified as W
 import Vira.Web.Widgets.Status qualified as Status
 import Vira.Web.Widgets.Time qualified as Time
 import Web.TablerIcons.Outline qualified as Icon
@@ -338,9 +339,9 @@ viraPRDetailsRow_ showRepo details = do
             PRNeverBuilt -> do
               -- No job, no unapproved: show PR title + state
               div_ [class_ "lg:col-span-8 flex items-center gap-2 flex-wrap text-sm"] $ do
-                prStateBadge_ pr.prState
+                W.prStateBadge_ pr.prState
                 span_ [class_ "text-gray-700 dark:text-gray-300 truncate max-w-md"] $ toHtml pr.title
-                forkBadge_ pr
+                W.forkBadge_ pr
               div_ [class_ "lg:col-span-4"] mempty
             PRBuilt job _freshness -> do
               -- Built: show job commit info + status
@@ -365,27 +366,3 @@ viraPRDetailsRow_ showRepo details = do
                   Just endTime -> Time.viraDuration_ $ diffUTCTime endTime job.jobCreatedTime
                   Nothing -> mempty
                 Status.viraStatusBadge_ job.jobStatus
-
--- | Badge for PR state (inlined to avoid circular import with PullPage)
-prStateBadge_ :: (Monad m) => PRState -> HtmlT m ()
-prStateBadge_ prState = do
-  let (colorClass, label) = case prState of
-        PROpen -> ("bg-green-100 dark:bg-green-900/30 text-green-800 dark:text-green-300", "Open" :: Text)
-        PRClosed -> ("bg-red-100 dark:bg-red-900/30 text-red-800 dark:text-red-300", "Closed")
-        PRMerged -> ("bg-purple-100 dark:bg-purple-900/30 text-purple-800 dark:text-purple-300", "Merged")
-      icon = case prState of
-        PROpen -> Icon.git_pull_request
-        PRClosed -> Icon.git_pull_request_closed
-        PRMerged -> Icon.git_merge
-  span_ [class_ $ "inline-flex items-center px-2 py-0.5 rounded-full text-xs font-medium " <> colorClass] $ do
-    div_ [class_ "w-3 h-3 mr-1 flex items-center justify-center"] $ toHtmlRaw icon
-    toHtml label
-
--- | Fork indicator badge (inlined to avoid circular import with PullPage)
-forkBadge_ :: (Monad m) => PullRequest -> HtmlT m ()
-forkBadge_ pr
-  | St.prIsFork pr =
-      span_ [class_ "inline-flex items-center px-2 py-0.5 rounded text-xs font-medium bg-purple-100 dark:bg-purple-900/30 text-purple-700 dark:text-purple-300"] $ do
-        div_ [class_ "w-3 h-3 mr-1 flex items-center justify-center"] $ toHtmlRaw Icon.git_fork
-        toHtml $ "fork: " <> unOwnerName pr.headOwner
-  | otherwise = mempty
